requirements.txt
torch>=1.12.0
torchvision>=0.13.0
omegaconf>=2.1.0
pillow>=9.0.0
tqdm>=4.64.0
numpy>=1.21.0
open-clip-torch>=2.0.0
opencv-python>=4.5.0
insightface>=0.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
